+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `kbgp_click_released` and `kbgp_user_action_released` for activating clicks
+  and user actions after the button was released.
 
 ## 0.12.0 - 2023-03-08
 ### Changed